% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicANCOVA.R
\name{ANCOVAplot}
\alias{ANCOVAplot}
\title{One-way ANCOVA Judgement and Plot}
\usage{
ANCOVAplot(x, y, groups, data, col = 1:length(levels(factor(groups))),
  pch = 1:length(levels(factor(groups))),
  lty = 1:length(levels(factor(groups))), Fig.slope = 1,
  legendPos = "topleft", ...)
}
\description{
One-way ANCOVA Judgement and Plot
}
\examples{
library(basicANCOVA)
data("isotope",package = "basicANCOVA")
View(isotope)
ANCOVAplot(x=isotope$d13C,y=isotope$d15N,groups=isotope$area,data=isotope)
}
