% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicANCOVA.R
\name{ANCOVAplot}
\alias{ANCOVAplot}
\title{One-way ANCOVA Judgement and Plot}
\usage{
ANCOVAplot(x, y, groups, data, col = 1:length(levels(factor(groups))),
  pch = 1:length(levels(factor(groups))),
  lty = 1:length(levels(factor(groups))), Fig.slope = 1, xlab = NULL,
  ylab = NULL, legendPos = "topleft", ...)
}
\arguments{
\item{x, y, groups}{For covariate 'x', dependent 'y' and factor 'groups'.}

\item{data}{The dataset contains three columns (x,y,groups) data. \cr See an example via data("isotope",package = "basicANCOVA").}

\item{col, pch, lty}{The color, pch and linetype for plot.}

\item{Fig.slope}{If 'Fig.slope = 1', draw graph and output the result of 'ANCOVA with same slope'. \cr  Whereas 'Fig.slope = 0', draw graph and output result of 'linear regression line' for each group.}

\item{xlab, ylab}{The labels of x-axis and y-axis.}

\item{legendPos}{The position of legend, such as one of c("none","bottomright","bottom","bottomleft","left","topleft","top","topright","right","center").}

\item{...}{additional parameters to \code{\link[graphics]{plot}},such as main, sub, xlab, ylab.}
}
\description{
One-way ANCOVA Judgement and Plot
}
\examples{

library(basicANCOVA)
data("hotdog",package = "HH")
View(isotope)

ANCOVAplot(x=hotdog$Sodium, y=hotdog$Calories, groups=hotdog$Type,
           data=hotdog, Fig.slope = 1) # (same slope) output ANCOVA results and graphic

# Not run
ANCOVAplot(x=hotdog$Sodium, y=hotdog$Calories, groups=hotdog$Type,
           data=isotope, Fig.slope = 0) # (different slopes) output "linear regression" results of each group and graphic
# End (Not run)

}
\seealso{
\code{\link[basicTrendline]{trendline}},  \code{\link[ancovaplot]{HH}}
}
\author{
Weiping Mei
}
